pyproject.toml
[tool.poetry]
name = "inventory_started"
version = "0.1.0"
description = ""
authors = ["Michele Costa <micosta@redhat.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"ruamel.yaml" = "^0.17.21"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
black = "^22.6.0"
ipdb = "^0.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
